- support for JSON-serialized tree transport - separated syntax/parsing and logic from visualization - added tests - fixed several bugs
