+++ .gitignore
@@ -1,2 +1,3 @@
 .gitconfig
 web/**
+node_modules/
