+++ btree.js
@@ -1,31 +1,27 @@
 //@ts-check
 
+// Node states
 const FAILED = 0,
       SUCCESS = 1,
       RUNNING = 2;
 
-const SAMPLE_TREE = `
-?
-|    ->
-|    |    (Ghost Close)
-|    |    ?
-|    |    |    ->
-|    |    |    |    !(Ghost Scared)
-|    |    |    |    (Power Pill Close)
-|    |    |    |    [Eat Power Pill]
-|    |    |    ->
-|    |    |    |    (Ghost Scared)
-|    |    |    |    [Chase Ghost]
-|    |    |    [Avoid Ghost]
-|    =1
-|    |    [Eat Pills]
-|    |    [Eat Fruit]
-`;
+// Node kinds
+const FALLBACK = 'fallback',
+      SEQUENCE = 'sequence',
+      PARALLEL = 'parallel',
+      ACTION = 'action',
+      CONDITION = 'condition';
 
 function expect(what, have) {
     return `Expecting '${what}', have '${have}'`;
 }
 
+/**
+ * Parses sequence node
+ * @param {string} buf behavior tree model
+ * @param {number} i current index
+ * @returns {[number, string | null]} tuple with adjusted current parsing index and error message or null
+ */
 function parseSequence(buf, i) {
     if (i < buf.length) {
         let ch = buf[i];
@@ -69,22 +65,22 @@ function parseAction(buf, i) {
 }
 
 function parseParallel(buf, i) {
-    let num = '';
+    let numBuf = '';
     while (i < buf.length) {
         let ch = buf[i];
         let m = ch.match(/\d/);
         if (m && m.length == 1) {
-            num += ch;
+            numBuf += ch;
         } else {
             break;
         }
         i++;
     }
-    if (num == '') {
+    if (numBuf === '') {
         return [0, i, 'Expecting number after parallel node.'];
     }
-    num = parseInt(num);
-    if (num == 0) {
+    let num = parseInt(numBuf);
+    if (num === 0) {
         return [0, i, 'Parallel node must allow at least one child.'];
     }
     return [num, i, null];
@@ -126,7 +122,7 @@ function node(name, kind, kids) {
 }
 
 function fallback() {
-    let fb = node('?', 'fallback', []);    
+    let fb = node('?', FALLBACK, []);
     fb.tick = function() {
         fb.active = true;
         for (let i = 0; i < fb.children.length; i++) {
@@ -143,7 +139,7 @@ function fallback() {
 }
 
 function sequence() {
-    let seq = node('\u2192', 'sequence', []);
+    let seq = node('\u2192', SEQUENCE, []);
     seq.tick = function() {
         seq.active = true;
         for (let i = 0; i < seq.children.length; i++) {
@@ -160,7 +156,8 @@ function sequence() {
 }
 
 function parallel(successCount) {
-    let par = node('\u21C9', 'parallel', []);
+    let par = node('\u21C9', PARALLEL, []);
+    par.successCount = successCount;
     par.tick = function() {
         par.active = true;
 
@@ -191,28 +188,107 @@ function parallel(successCount) {
 }
 
 function action(name) {
-    let a = node(name, 'action');
+    let a = node(name, ACTION);
     a.setStatus(RUNNING);
     return a;
 }
 
-function condition(name) { return node(name, 'condition'); }
+function condition(name, hasNot) {
+    let c = node(name, CONDITION);
+    c.hasNot = hasNot; // ensure the property is declared in both cases
+    return c;
+}
 
+class BehaviorTree {
+    /**
+     * Behavior Tree
+     * @param {any} root tree root node
+     * @param {Map<string, any[]>} actions list of actions grouped by name
+     * @param {Map<string, any[]>} conditions list of conditions grouped by name
+     * @param {number} line line at which the error ocurred
+     * @param {string} error parsing error
+     */
+    constructor(root, actions, conditions, line, error = null) {
+        this.root = root;
+        this.actions = actions;
+        this.conditions = conditions;
+        this.line = line;
+        this.error = error;
+    }
+
+    /**
+     * Re-builds the behavior tree from JSON e.g. after it has been loaded from a file, or transferred via http.
+     * @param {any} treeAsJson behavior tree in a JSON form
+     * @returns {BehaviorTree}
+     */
+    static fromJson(treeAsJson) {
+        let actions = new Map();
+        let conditions = new Map();
+        let rootNode = BehaviorTree.nodeFromJson(treeAsJson.root, actions, conditions);
+        return new BehaviorTree(rootNode, actions, conditions, treeAsJson.line, treeAsJson.error);
+    }
+
+    /**
+     * Re-builds tree node.
+     * @param {any} nodeAsJson node in plain JSON form
+     * @param {Map<string, any[]>} actions map to register all actions found
+     * @param {Map<string, any[]>} conditions map to register all conditions found
+     */
+    static nodeFromJson(nodeAsJson, actions, conditions) {
+        let node;
+        switch (nodeAsJson.kind) {
+            case FALLBACK:
+                node = fallback();
+                break;
+            case SEQUENCE:
+                node = sequence();
+                break;
+            case PARALLEL:
+                node = parallel(nodeAsJson.successCount);
+                break;
+            case ACTION:
+                node = action(nodeAsJson.name);
+                addToArrayMap(actions, node.name, node);
+                break;
+            case CONDITION:
+                node = condition(nodeAsJson.name, nodeAsJson.hasNot);
+                addToArrayMap(conditions, node.name, node);
+                break;
+            default:
+                throw new Error(`Unexpected node kind: ${nodeAsJson.kind}.`);
+        }
+
+        if (nodeAsJson.children) {
+            node.children = nodeAsJson.children.map(child => this.nodeFromJson(child, actions, conditions));
+        }
+
+        return node;
+    }
+}
+
+/**
+ * Parser
+ * @param {string} buf behavior tree as text
+ * @returns {BehaviorTree}
+ */
 function parse(buf) {
-    let indent     = 0,      // current recorded indentation
-        line       = 1,      // line number in text
-        nodes      = [null], // node tree
-        actions    = {},     // unique action nodes
-        conditions = {},     // unique condition nodes
-        notPending = false,  // is 'not' decorator waiting to be applied?
+    let indent     = 0,        // current recorded indentation
+        line       = 1,        // line number in text
+        nodes      = [null],   // node tree
+        actions    = new Map(),// action nodes grouped by name
+        conditions = new Map(),// condition nodes grouped by name
+        notPending = false,    // is 'not' decorator waiting to be applied?
         i          = 0;
 
     function pushNode(node) {
         if (indent === 0 && nodes[indent]) {
-            return 'More than one root node';
+            return `More than one root node or node '${node.name}' has wrong indentation.`;
         }
-        let parent = nodes[indent - 1];
-        if (parent) {
+        if (indent > 0) {
+            let parent = nodes[indent - 1];
+            if (!parent) {
+                return `${node.name} node has no parent (wrong indentation level)`;
+            }
             if (parent.children) {
                 parent.children.push(node);
                 nodes[indent] = node;
@@ -222,12 +298,12 @@ function parse(buf) {
         } else {
             nodes[indent] = node;
         }
-        indent = 0;
+        indent++; // nested child on the same line should be indented
         return null;
     };
 
     function onError(err) {
-        return {root: null, actions, conditions, line, error: err};
+        return new BehaviorTree(null, actions, conditions, line, err);
     }
 
     while (i < buf.length) {
@@ -241,14 +317,16 @@ function parse(buf) {
             break;
 
         case '\r': {
-            if (i < buf.length &&  buf[i] === '\n') {
+            if (i < buf.length && buf[i] === '\n') {
                 i += 1;
             }
             line++;
+            indent = 0;
         } break;
 
         case '\n': {
             line++;
+            indent = 0;
         } break;
 
         case '|': {
@@ -288,11 +366,11 @@ function parse(buf) {
         case '-': {
             let [n, err] = parseSequence(buf, i);
             if (err) {
-                return {root: null, actions, conditions, line, err};
+                return onError(err);
             }
             i = n;
             let e = pushNode(sequence());
-            if (err) {
+            if (e) {
                 return onError(e);
             }
         } break;
@@ -303,15 +381,10 @@ function parse(buf) {
                 return onError(err);
             }
             i = n;
-            let c = condition(name);
-            if (conditions[name]) {
-                conditions[name].push(c);
-            } else {
-                conditions[name] = [c];
-            }
+            let c = condition(name, notPending);
+            addToArrayMap(conditions, name, c);
             if (notPending) {
                 notPending = false;
-                c.hasNot = true;
             }
             let e = pushNode(c);
             if (e) {
@@ -326,11 +399,7 @@ function parse(buf) {
             }
             i = n;
             let a = action(name);
-            if (actions[name]) {
-                actions[name].push(a);
-            } else {
-                actions[name] = [a];
-            }
+            addToArrayMap(actions, name, a);
             let e = pushNode(a);
             if (e) {
                 return onError(e);
@@ -353,420 +422,41 @@ function parse(buf) {
         return onError(e);
     }
 
-    return { root: nodes[0], line, actions, conditions, error: null };
+    return new BehaviorTree(nodes[0], actions, conditions, line, null);
 }
 
-// Render _root_ inside the _parent_ DOM node with a viewbox width
-// of _width_.  _x0_ and _x1_ are used to determine the vertical height
-// of the tree inside the viewbox.
-function renderTree(parent, root, width, x0, x1) {
-    function translate(tree) {
-        let x = tree.dy + tree.drag_dx,
-            y = tree.dx - x0 + tree.drag_dy;
-        return `translate(${x}, ${y}) scale(${root.scale})`;
-    }
-
-    const svg = d3.select(parent)
-          .html('')
-          .append('svg')
-          .attr('width', '100%')
-          .attr('height', '100%')
-          .attr('viewBox', [0, 0, width, x1 - x0 + root.dx * 4]);
-    
-    const g = svg.append('g')
-          .attr('font-family', 'sans-serif')
-          .attr('font-size', 12)
-          .attr('transform', translate(root));
-
-    svg.call(d3.drag().on('drag.svg', function() {
-        let s = clamp(root.scale, 1.0, root.scale);
-        root.drag_dy += d3.event.dy * s;
-        root.drag_dx += d3.event.dx * s;
-        g.attr('transform', translate(root));
-    }));
-    
-    const link = g.append('g')
-          .attr('fill', 'none')
-          .attr('stroke', '#555')
-          .attr('stroke-opacity', 0.4)
-          .attr('stroke-width', 1.5)
-          .selectAll('path')
-          .data(root.links())
-          .join('path')
-          .attr('d', d3.linkHorizontal()
-                .x(d => d.y)
-                .y(d => d.x));
-    
-    const node = g.append('g')
-          .attr('stroke-linejoin', 'round')
-          .attr('stroke-width', 3)
-          .selectAll('g')
-          .data(root.descendants())
-          .join('g')
-          .attr('transform', d => `translate(${d.y},${d.x})`);
-
-    function nodeColor(active, status) {
-        let base       = 'BF',
-            amp        = '11',
-            color      = '#${base}${amp}${amp}',
-            fill       = 'white',
-            text_color = 'black';
-
-        if (active) {
-            amp = '50';
-        }
-        switch (status) {
-        case FAILED:  color = `#${base}${amp}${amp}`; break;
-        case SUCCESS: color = `#${amp}${base}${amp}`; break;
-        case RUNNING: color = `#${amp}${amp}${base}`; break;
-        }
-        return color;
-    }
-
-    node.each(function(d) {
-        let active = d.data.active,
-            k      = d.data.kind;
-
-        if (k == 'sequence' || k == 'fallback' || k == 'parallel') {
-            let color      = nodeColor(active, d.data.status()),
-                fill       = 'white',
-                text_color = 'black';
-
-            if (active) {
-                fill = color;
-                color = '#444';
-                text_color = 'white';
-            }
-
-            const SZ = 24;
-
-            d3.select(this)
-                .append('rect')
-                .attr('x', -SZ / 2)
-                .attr('y', -SZ / 2)
-                .attr('width', SZ)
-                .attr('height', SZ)
-                .attr('fill', fill)
-                .attr('stroke-width', 2)
-                .attr('stroke', color);
-            d3.select(this)
-                .append('text')
-                .attr('dy', '0.31em')
-                .attr('x', d => d.data.kind == 'sequence' ? 5 : 3)
-                .attr('text-anchor', 'end')
-                .text(d.data.name)
-                .attr('fill', text_color)
-                .clone(true).lower();
-        }
-
-        if (k == 'condition' || k == 'action') {
-            let container,
-                color      = nodeColor(active, d.data.status()),
-                fill       = 'white',
-                text_color = 'black';
-
-            if (active) {
-                fill = color;
-                color = '#111';
-                text_color = 'white';
-            }
-
-            if (k == 'condition') {
-                container = d3.select(this).append('ellipse');
-            }
-            if (k == 'action') {
-                container = d3.select(this).append('rect');
-            }
-            container
-                .attr('fill', fill)
-                .attr('stroke-width', 2)
-                .attr('stroke', color);
-            
-            const PAD = 10;
-
-            let name = d.data.name;
-            if (d.data.hasNot) {
-                name = '!' + name;
-            }
-            let text = d3.select(this)
-                .append('text')
-                .attr('dy', '0.31em')
-                .attr('text-anchor', 'start')
-                .attr('fill', text_color)
-                .text(name)
-                .clone(true).lower()
-                .node();
-
-            let width = text.getComputedTextLength() + PAD;
-            if (k == 'condition') {
-                container
-                    .attr('cx', width/2.0 - PAD/2.0)
-                    .attr('rx', width/1.75)
-                    .attr('ry', '1.0em');
-            }
-            if (k == 'action') {
-                container
-                    .attr('y', '-0.85em')
-                    .attr('x', -PAD/2.0)
-                    .attr('width', width)
-                    .attr('height', '1.75em');
-            }
-        }
-
-        // node tooltip
-        var status = getFriendlyStatus(d.data.status());
-        d3.select(this)
-            .append("svg:title")
-            .text(d => `Node: ${d.data.hasNot?"NOT ": ""}${d.data.name} ${k}\nActive: ${active}\nStatus: ${status}`);
-    });
-}
-
-/**
- * Gets friendly status
- * @param {number} status tree node status
- * @returns {string} user-friendly status string
- */
-function getFriendlyStatus(status) {
-    switch (status) {
-        case FAILED:
-            return 'Failed';
-        case SUCCESS:
-            return 'Success';
-        case RUNNING:
-            return 'Running';
-        default:
-            return 'Unknown';
+function addToArrayMap(map, key, value) {
+    if (map[key]) {
+        map[key].push(value);
+    } else {
+        map[key] = [value];
     }
 }
 
-function clamp(val, min, max) {
-    if (val < min) {
-        return min;
-    }
-    if (val > max) {
-        return max;
-    }
-    return val;
+const SAMPLE_TREE = `
+?
+|    ->
+|    |    (Ghost Close)
+|    |    ?
+|    |    |    ->
+|    |    |    |    !(Ghost Scared)
+|    |    |    |    (Power Pill Close)
+|    |    |    |    [Eat Power Pill]
+|    |    |    ->
+|    |    |    |    (Ghost Scared)
+|    |    |    |    [Chase Ghost]
+|    |    |    [Avoid Ghost]
+|    =1
+|    |    [Eat Pills]
+|    |    [Eat Fruit]
+`;
+
+if (typeof exports !== 'undefined' && exports) {
+    exports.bt = {
+        BehaviorTree,
+        parse, SUCCESS, FAILED, RUNNING,
+        fallback, sequence, parallel, condition, action,
+        FALLBACK, SEQUENCE, PARALLEL, CONDITION, ACTION,
+        SAMPLE_TREE
+    };
 }
-
-function windowSize() {
-    let w = window,
-        d = document,
-        e = d.documentElement,
-        g = d.getElementsByTagName('body')[0],
-        x = w.innerWidth || e.clientWidth || g.clientWidth,
-        y = w.innerHeight || e.clientHeight || g.clientHeight;
-    return [x, y];
-}
-
-// LoadTree parses _str_ into a root tree node and renders it.
-function loadTree(str) {
-    let result = parse(str),
-        line   = result.line;
-    if (result.error) {
-        d3.select('main')
-            .html('')
-            .append('div')
-            .classed('tree-error__container', true)
-            .append('span')
-            .classed('tree-error__text', true)
-            .text(`Line ${line}: ${result.error}`);
-        return;
-    }
-
-    let x0   = Infinity,
-        x1   = -x0,
-        data = d3.hierarchy(result.root),
-        root = undefined,
-        horizontal_stretch = 0,
-        vertical_stretch   = 8,
-        [width, height]    = windowSize();
-
-    data.drag_dx = 0;
-    data.drag_dy = 0;
-    data.scale = 1;
-
-    function resizeRoot() {
-        data.dx = vertical_stretch;
-        data.dy = (width + horizontal_stretch) / (data.height + 3);
-
-        root = d3.tree().nodeSize([4*data.dx, data.dy])(data);
-        root.each(d => {
-            if (d.x > x1) x1 = d.x;
-            if (d.x < x0) x0 = d.x;
-        });
-    }
-    resizeRoot();
-
-    window.addEventListener('resize', function() {
-        [width, height] = windowSize();
-        resizeRoot();
-        render();
-    });
-
-    d3.select('main')
-        .on('wheel', function() {
-            let e = d3.event;
-            let up = false;
-
-            if (e.deltaY == 0) {
-                return;
-            }
-            if (e.deltaY < 0) {
-                up = true;
-            }
-            if (e.shiftKey && e.altKey) {
-                vertical_stretch += up ? 1 : -1;
-                vertical_stretch = clamp(vertical_stretch, 3, 1000);
-            } else if (e.shiftKey) {
-                horizontal_stretch += up ? 30 : -30;
-                horizontal_stretch = clamp(horizontal_stretch, -width*0.6, 1000.0);
-            } else {
-                data.scale += up ? 0.10 : -0.10;
-                data.scale = clamp(data.scale, 0.10, 100.0);
-            }
-            resizeRoot();
-            render();
-        });
-
-    let conds = d3.select('#tree-conditions')
-        .html('')
-        .selectAll('a')
-        .data(Object.keys(result.conditions).sort())
-        .enter()
-        .append('a')
-        .classed('mdl-navigation__link', true);
-
-    let condLabels = conds.append('label')
-        .attr('for', function(d, i) { return `switch-${i}`; });
-    
-    condLabels.append('input')
-        .attr('type', 'checkbox')
-        .attr('id', function(d, i) { return `switch-${i}`; })
-        .classed('mdl-switch__input', true)
-        .on('change', function(name) {
-            let s = d3.event.target.checked ? SUCCESS : FAILED;
-            result.conditions[name].forEach(c => c.setStatus(s));
-            render();
-        });
-    condLabels
-        .classed('mdl-switch mdl-js-switch mdl-js-ripple-effect', true);
-    condLabels.append('span')
-        .classed('mdl-switch__label', true)
-        .text(name => name);
-    // Force the material library to call the JS on all label
-    // elements; otherwise, if loading a new tree the switches will
-    // appear as checkboxes.
-    condLabels.each(function(d) {
-        let label = d3.select(this).node();
-        componentHandler.upgradeElement(label);
-    });
-
-    const BTN_CLASS = 'mdl-button mdl-js-button mdl-button--fab tree-action--mini-fab mdl-js-ripple-effect';
-
-    let actions = d3.select('#tree-actions')
-        .html('')
-        .selectAll('a')
-        .data(Object.keys(result.actions).sort())
-        .enter()
-        .append('a')
-        .classed('mdl-navigation__link tree-action', true);
-    actions.append('span')
-        .text(name => name);
-
-    let actionBtns = actions.append('div');
-
-    let clear = actionBtns.append('button')
-        .classed(BTN_CLASS, true)
-        .classed('tree-action--failure', true)
-        .attr('title', name => "Set action '" + name + "' as failed")
-        .on('click', function(name) {
-            result.actions[name].forEach(a => a.setStatus(FAILED));
-            render();
-        });
-    clear
-        .append('i')
-        .classed('material-icons', true)
-        .text('clear');
-    clear.each(function(d) {
-        let btn = d3.select(this).node();
-        componentHandler.upgradeElement(btn);
-    });
-
-    let add = actionBtns.append('button')
-        .classed(BTN_CLASS, true)
-        .classed('tree-action--success', true)
-        .attr('title', name => "Set action '" + name + "' as succeeded")
-        .on('click', function(name) {
-            result.actions[name].forEach(a => a.setStatus(SUCCESS));
-            render();
-        });
-    add
-        .append('i')
-        .classed('material-icons', true)
-        .text('add');
-    add.each(function(d) {
-        let btn = d3.select(this).node();
-        componentHandler.upgradeElement(btn);
-    });
-
-    let run = actionBtns.append('button')
-        .classed(BTN_CLASS, true)
-        .classed('tree-action--running', true)
-        .attr('title', name => "Start '" + name + "' action")
-        .on('click', function(name) {
-            result.actions[name].forEach(a => a.setStatus(RUNNING));
-            render();
-        });
-    run
-        .append('i')
-        .classed('material-icons', true)
-        .text('play_arrow');
-    run.each(function(d) {
-        let btn = d3.select(this).node();
-        componentHandler.upgradeElement(btn);
-    });
-
-    function render() {
-        root.data.deactivate();
-        root.data.tick();
-        renderTree('main', root, width, x0, x1);
-    }
-    render();
-}
-
-(function main() {
-    let treeSelect = document.getElementById('treeFileSelect'),
-        treeInput  = document.getElementById('treeFileInput');
-
-    treeInput.addEventListener('change', function() {
-        if (this.files.length < 1) {
-            return;
-        }
-        let reader = new FileReader();
-        reader.onload = function(e) {
-            loadTree(e.target.result);
-        };
-        reader.readAsText(this.files[0]);
-    }, false);
-
-    treeSelect.addEventListener('click', function(e) {
-        if (treeInput) {
-            treeInput.click();
-        }
-    }, false);
-
-    d3.select('#tree-help__button')
-        .on('click', function() {
-            let card = d3.select('#tree-help__card'),
-                viz  = card.style('visibility');
-            if (viz == 'hidden') {
-                viz = 'visible';
-            } else {
-                viz = 'hidden';
-            }
-            card.style('visibility', viz);
-        });
-
-    loadTree(SAMPLE_TREE);
-})();
