+++ package.json
@@ -1,0 +1,14 @@
+{
+    "scripts": {
+        "test": "mocha"
+    },
+    "dependencies": {
+        "body-parser": "^1.19.0",
+        "cors": "^2.8.5",
+        "express": "^4.17.1"
+    },
+    "devDependencies": {
+        "chai": "^4.2.0",
+        "mocha": "^7.1.0"
+    }
+}
